Interpret windows headers in either 32- or 64-bit.
Windows uses the same header files on both 32- and 64-bit
machines. They are simply interpreted differently. For example,
a ULONG_PTR value is 32 bits when complied with a 32-bit compiler and
is 64 bits when compiled with a 64-bit compiler. Therefore, CParser
also needs to be able to interpret the header files in 32-bit mode
or 64-bit mode.

This commit adds code for determining which mode to use and giving
CParser the appropriate starting macros. It also adds a separate 32 bit
cache and 64 bit cache.

(cherry picked from commit f0252bcbb7e2ed9e8b9565300ab4ec40fe02075b)